requirements.txt
alembic==1.4.2
astroid==2.2.5
Click==7.0
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==20.0.4
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycryptodome==3.3.1
pylint==2.3.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
python-jose-cryptodome==1.3.2
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
Werkzeug==0.15.2
wrapt==1.11.1
